requirements.txt
numpy==1.24.4
Pillow==11.0.0
streamlit==1.39.0
tensorflow==2.9.3
